Refactor: add a new file to keep functions for common user
